toggleSettings Map, Refactor, Reformat
Added toggleSettings Mapping and a getter for it.

Refactored the ConfigValidator.getMapping() to ConfigValidator.getLocaleMap()

Moved the loading of Maps in ConfigValidator to their own methods.

Hit CTRL+ALT+L on all files to reformat.

**NOTE**
ConfigValidator.getToggleSettings() is unused right now, none of the current config calls have been replaced for that in VillagerHandler.